Reworked root search in BasicHierarchy convenience constructor
Now expects the nodes list to be sorted, and simply takes the first node as root.
This change was necessitated by a use case in which a BasicHierarchy object was created around an existing Node from another hierarchy, to create a sub-hierarchy. Since the Node already had a parent, the constructor could not find a root and failed.